\name{rmutil}
\title{Utilities for Repeated Measurements Library}
\description{
\code{\link{dbetabinom}}		Density of Beta Binomial Distribution

\code{\link{dboxcox}}			Density of Box-Cox Distribution

\code{\link{dburr}}			Density of Burr Distribution

\code{\link{ddoublebinom}}		Density of Double Binomial Distribution

\code{\link{ddoublepois}}		Density of Double Poisson Distribution

\code{\link{dgammacount}}		Density of Gamma Count Distribution

\code{\link{dgextval}}		Density of Generalized Extreme Value Distribution

\code{\link{dggamma}}			Density of Generalized Gamma Distribution

\code{\link{dginvgauss}}		Density of Generalized Inverse Gaussian Distribution

\code{\link{dglogis}}			Density of Generalized Logistic Distribution

\code{\link{dgweibull}}		Density of Generalized Weibull Distribution

\code{\link{dhjorth}}			Density of Hjorth Distribution

\code{\link{dinvgauss}}		Density of Inverse Gaussian Distribution

\code{\link{dlaplace}}		Density of Laplace Distribution

\code{\link{dlevy}}			Density of Levy Distribution

\code{\link{dmultbinom}}		Density of Multiplicative Binomial Distribution

\code{\link{dmultpois}}		Density of Multiplicative Poisson Distribution

\code{\link{dpareto}}			Density of Pareto Distribution

\code{\link{dpowexp}}			Density of Power Exponential Distribution

\code{\link{finterp}}			Formula Interpreter

\code{\link{fnenvir}}			Check Covariates and Parameters of a Function

\code{\link{gauss.hermite}}	Calculate Gauss-Hermite Quadrature Points

\code{\link{gettvc}}			Create Time-varying Covariates

\code{\link{int}}			Numerical Integration

\code{\link{mexp}}			Matrix Exponentiation

\code{\link{pbetabinom}}		Distribution Function of Beta Binomial Distribution

\code{\link{pboxcox}}			Distribution Function of Box-Cox Distribution

\code{\link{pburr}}			Distribution Function of Burr Distribution

\code{\link{pdoublebinom}}		Distribution Function of Double Binomial Distribution

\code{\link{pdoublepois}}		Distribution Function of Double Poisson Distribution

\code{\link{pgammacount}}		Distribution Function of Gamma Count Distribution

\code{\link{pgextval}}		Distribution Function of Generalized Extreme Value Distribution

\code{\link{pggamma}}			Distribution Function of Generalized Gamma Distribution

\code{\link{pginvgauss}}		Distribution Function of Generalized Inverse Gaussian Distribution

\code{\link{pglogis}}			Distribution Function of Generalized Logistic Distribution

\code{\link{pgweibull}}		Distribution Function of Generalized Weibull Distribution

\code{\link{phjorth}}			Distribution Function of Hjorth Distribution

\code{\link{pinvgauss}}		Distribution Function of Inverse Gaussian Distribution

\code{\link{pkpd}}			Pharmacokinetic Model Functions

\code{\link{plaplace}}		Distribution Function of Laplace Distribution

\code{\link{plevy}}			Distribution Function of Levy Distribution

\code{\link{plot.iprofile}}	Plot Individual Time Profiles for Carma

\code{\link{plot.profile}}		Plot Marginal Time Profiles for Carma

\code{\link{plot.residuals}}	Plot Residuals for Carma

\code{\link{pmultbinom}}		Distribution Function of Multiplicative Binomial Distribution

\code{\link{pmultpois}}		Distribution Function of Multiplicative Poisson Distribution

\code{\link{ppareto}}			Distribution Function of Pareto Distribution

\code{\link{ppowexp}}			Distribution Function of Power Exponential Distribution

\code{\link{read.list}}		Read a List of Matrices of Unbalanced Repeated Measurements from a File

\code{\link{read.surv}}		Read a List of Vectors of Event Histories from a File

\code{\link{rmna}}			Create a Repeated Object

\code{\link{restovec}}		Create a Response Object

\code{\link{tcctomat}}		Create a Time-constant Covariate (tccov) Object

\code{\link{tvctomat}}		Create a Time-varying Covariate (tvcov) Object

\code{\link{wr}}				Find the Response Vector and Design Matrix for a Model Formula
}
